src/requirements.txt
Wave>=0.0.2
python-memcached>=1.50
requests>=2.4.3
wsgiref>=0.1.2
CherryPy>=4.0.0
python-vlc>=1.1.2
pyalsaaudio>=0.8.2
py-getch>=1.0.1
pocketsphinx>=0.1.3
webrtcvad>=2.0.7
hyper>=0.7.0
RPi.GPIO
